Moved django_adapter to outisde RIP and into django-RIP
