+++ .travis.yml
@@ -3,7 +3,6 @@
 language: python
 
 python:
-
   - "2.7"
   - "pypy"
