+++ django_adapter/action_resolver.py
