+++ django_adapter/metadata_factory.py
