+++ tests/integration_tests/test_crud_resource_end_to_end.py
@@ -6,10 +6,10 @@ from hamcrest.core.assert_that import assert_that
 from hamcrest.core.core.isequal import equal_to
 from mock import MagicMock
 
+from django_adapter import create_http_handler
 from rip.api import Api
 from rip.api_schema import ApiSchema
 from rip.crud.crud_resource import CrudResource
-from rip.django_adapter.django_http_handler import create_http_handler
 from rip.generic_steps.default_authentication import DefaultAuthentication
 from rip.generic_steps.default_authorization import DefaultAuthorization
 from rip.generic_steps.default_entity_actions import DefaultEntityActions
